# Avoid extra allocation and interface calls in Append