Small tweak to Enumerable.Append to avoid an unnecessary enumerator allocation, avoid iterating through that enumerator via an interface, and avoid an unnecessary method on a generic class.

Before

|  Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------- |----------:|----------:|----------:|-------:|------:|------:|----------:|
|     One |  26.35 ns | 0.2571 ns | 0.2279 ns | 0.0102 |     - |     - |      64 B |
|     Ten |  90.30 ns | 0.3271 ns | 0.2900 ns | 0.0305 |     - |     - |     192 B |
| Hundred | 589.73 ns | 5.5760 ns | 5.2158 ns | 0.1450 |     - |     - |     912 B |


After

|  Method |      Mean |     Error |    StdDev |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------- |----------:|----------:|----------:|-------:|-------:|------:|----------:|
|     One |  25.85 ns | 0.2254 ns | 0.2108 ns | 0.0102 |      - |     - |      64 B |
|     Ten |  66.58 ns | 0.6744 ns | 0.5979 ns | 0.0254 |      - |     - |     160 B |
| Hundred | 209.70 ns | 1.0881 ns | 1.0178 ns | 0.1402 | 0.0005 |     - |     880 B |


Benchmark
```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;
using System.Linq;
using System.Collections.Generic;

[MemoryDiagnoser]
public class Program
{
    static void Main(string[] args) => BenchmarkSwitcher.FromTypes(new[] { typeof(Program) }).Run(args);

    private IEnumerable<int> _one;
    private IEnumerable<int> _ten;
    private IEnumerable<int> _hundred;

    [GlobalSetup]
    public void GlobalSetup()
    {
        _one = Enumerable.Empty<int>().Append(1);

        _ten = Enumerable.Empty<int>();
        for (int i = 0; i < 10; i++) _ten = _ten.Append(i);

        _hundred = Enumerable.Empty<int>();
        for (int i = 0; i < 100; i++) _hundred = _hundred.Append(i);
    }

    [Benchmark] public List<int> One() => _one.ToList();
    [Benchmark] public List<int> Ten() => _ten.ToList();
    [Benchmark] public List<int> Hundred() => _hundred.ToList();
}
```

Per the discussion in https://github.com/dotnet/corefx/pull/38975#discussion_r298195901
cc: @timandy, @maryamariyan